staff_picks is now a thing.
 please remember to migrate;  Also, since POSTGRES always needs a query
order, I changed the Stream.js model to allow for passing in an empty
"collection options" hash, which if over-ridden, changes the stream
models default comparator, which was preventing us from sorting on a
different timestamp, i.e. such as like creation time, or
staff_picked_at.  as such, the like stream is now sorted by WHEN the
user liked it, rather than posts.created_at.  This should not change
the behavior for streams that do not override the collectionOptions
option, (the profile page)